NXPY-131: Make the HTTP response logging safer
Now, no excuses: we cannot have MemoryError
nor OverflowError anymore. The HTTP response
logger is safe to use.

I also added a simple module attribute
`skip_logging` to define at `True` in test files
where we do not need to log the current test on
the server. This will prevent having to create
a connection to the server for nothing.