NXPY-155: Don't use dots in custom HTTP headers
We don't want to use dots nor underscores in headers
sent to Nuxeo as NGINX doesn't allow their usage
for HTTP.
Underscores and dots have been replaced with hyphen to
not have the issue with NGINX users.
A warning is now thrown in case a dot or an underscore
is encountered in a header or a request params.
Tests will now fail if a warning is encountered.
Also changelog has been updated.