NXPY-171: Set higher default timeouts
- `TIMEOUT_CONNECT`: 5 -> 10
- `TIMEOUT_READ`: 30 -> 600

It happens that the `TIMEOUT_READ` is set at the socket level at the first HTTP(S) connection. And then this value is used for next calls. And even if we set a higher value in specific calls, it will not be taken into account. So we are now using a bigger value to prevent `socket.timeout: The write operation timed out` errors at the SSL level.